SNOW-1011765: Moved list-images and list-tags to from image-registry to image-repository. Added integration tests for image-repository
